separates bls12_381 into 7 files
