+++ actors/evm/src/interpreter/precompiles/bls12_381/mod.rs
@@ -1,0 +1,15 @@
+mod g1_add;
+mod g1_msm;
+mod g2_add;
+mod g2_msm;
+mod pairing;
+mod map_fp_to_g1;
+mod map_fp2_to_g2;
+
+pub use g1_add::bls12_g1add;
+pub use g1_msm::bls12_g1msm;
+pub use g2_add::bls12_g2add;
+pub use g2_msm::bls12_g2msm;
+pub use pairing::bls12_pairing;
+pub use map_fp_to_g1::bls12_map_fp_to_g1;
+pub use map_fp2_to_g2::bls12_map_fp2_to_g2;
