+++ actors/evm/src/interpreter/precompiles/bls12_381/pairing.rs
@@ -1,0 +1,24 @@
+use crate::interpreter::{
+    precompiles::{PrecompileContext, PrecompileError, PrecompileResult},
+    System,
+};
+use fil_actors_runtime::runtime::Runtime;
+
+/// BLS12_PAIRING precompile
+/// Implements BLS12-381 pairing check according to EIP-2537
+#[allow(dead_code,unused_variables)]
+pub fn bls12_pairing<RT: Runtime>(
+    _: &mut System<RT>,
+    input: &[u8],
+    _: PrecompileContext,
+) -> PrecompileResult {
+    Err(PrecompileError::CallForbidden)
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use crate::interpreter::System;
+    use fil_actors_runtime::test_utils::MockRuntime;
+    use hex_literal::hex;
+}
